Add files via upload
I have launched the browser and entered the URL (www.ahibus.com).
I landed on the Home Page.
I have selected From Station as Kumbakonam using the sendkeys. Following
that I have selected the To Station as Chennai. Then I selected the Date. Once the Search button is clicked, based on the Search the Travels page will
displayed. Then I selected the bus name and the required Seat by clicking on the Show
Seats option. System directs us to select the Boarding and Dropping point. I selected
Kumbakonam as the Boarding point and Kalpakkam as the Dropping point.
I was directed to the Login Pop-up screen when clicked on the Continue
button. Once the Login page is skipped, Passenger details page is displayed. I filled
the following details in the Passenger details page - Name, Age, Gender, Mobile number, and Email. Once the Passenger details are filled, I go to Payment page, to make payment.